requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
boto3==1.16.56
botocore==1.19.56
Django==3.1.5
django-extensions==3.1.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-storages==1.11.1
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.2.2
graphql-core==2.3.2
graphql-relay==2.0.1
jmespath==0.10.0
Pillow==8.1.0
promise==2.3
PyJWT==1.7.0
python-dateutil==2.8.1
pytz==2020.5
Rx==1.6.1
s3transfer==0.3.4
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
